relocate UPD/NTP functions to separate file
