+++ lib/HelperLibs/HelperFncs.h
@@ -6,9 +6,6 @@
 #include <ArduinoOTA.h>
 #include <WiFiUdp.h>
 
-ESP8266WebServer server(80);     // Create a webserver instance
-WiFiUDP UDP;                     // Create an instance of the WiFiUDP class to send and receive
-
 // change these values to match your network
 const char* ssid = "Vodafone-34321872";       //  your network SSID (name)
 const char* pass = "2smid42fpfix3pe";         // your network password
@@ -26,11 +23,6 @@ int calcHumidity = 0;
 float voltage = 0;
 
 size_t fileSizeTemp;
-
-IPAddress timeServerIP;          // ntp1.inrim.it NTP server address
-const char* NTPServerName = "ntp1.inrim.it";
-const int NTP_PACKET_SIZE = 48;  // NTP time stamp is in the first 48 bytes of the message
-byte NTPBuffer[NTP_PACKET_SIZE]; // buffer to hold incoming and outgoing packets
 
 //---------------------USEFUL FUNCTIONS-----------------------
 
@@ -99,6 +91,9 @@ void startWebServer() {
 }
 */
 
+
+
+
 //---------------------STARTUP FUNCTIONS-----------------------
 
 void startSerial() {
@@ -161,51 +156,4 @@ void startOTA() {
   });
   ArduinoOTA.begin();
   Serial.println("OTA ready");
-}
-
-void startUDP() {
-  Serial.println("Starting UDP");
-  UDP.begin(123);                          // Start listening for UDP messages on port 123
-  Serial.print("Local port:\t");
-  Serial.println(UDP.localPort());
-}
-
-//---------------------NTP FUNCTIONS-----------------------
-
-void sendNTPpacket(IPAddress& address) {
-  memset(NTPBuffer, 0, NTP_PACKET_SIZE);  // set all bytes in the buffer to 0
-  // Initialize values needed to form NTP request
-  NTPBuffer[0] = 0b11100011;   // LI, Version, Mode
-  // send a packet requesting a timestamp:
-  UDP.beginPacket(address, 123); // NTP requests are to port 123
-  UDP.write(NTPBuffer, NTP_PACKET_SIZE);
-  UDP.endPacket();
-}
-
-uint32_t getTime() {
-  if (UDP.parsePacket() == 0) { // If there's no response (yet)
-    return 0;
-  }
-  UDP.read(NTPBuffer, NTP_PACKET_SIZE); // read the packet into the buffer
-  // Combine the 4 timestamp bytes into one 32-bit number
-  uint32_t NTPTime = (NTPBuffer[40] << 24) | (NTPBuffer[41] << 16) | (NTPBuffer[42] << 8) | NTPBuffer[43];
-  // Convert NTP time to a UNIX timestamp:
-  // Unix time starts on Jan 1 1970. That's 2208988800 seconds in NTP time:
-  const uint32_t seventyYears = 2208988800UL;
-  // subtract seventy years:
-  uint32_t UNIXTime = NTPTime - seventyYears;
-  return UNIXTime;
-}
-
-void startNTP() {
-  if(!WiFi.hostByName(NTPServerName, timeServerIP)) { // Get the IP address of the NTP server
-    Serial.println("DNS lookup failed. Rebooting.");
-    Serial.flush();
-    ESP.reset();
-  }
-  Serial.print("Time server IP:\t");
-  Serial.println(timeServerIP);
-  
-  Serial.println("\r\nSending first NTP request ...");
-  sendNTPpacket(timeServerIP);
 }
